Fixes nameclature on on Repository
