+++ artemis-core/src/main/java/org/jnosql/artemis/column/query/ColumnCrudRepositoryAsyncProxy.java
@@ -16,7 +16,7 @@
 package org.jnosql.artemis.column.query;
 
 
-import org.jnosql.artemis.CrudRepositoryAsync;
+import org.jnosql.artemis.RepositoryAsync;
 import org.jnosql.artemis.DynamicQueryException;
 import org.jnosql.artemis.column.ColumnTemplateAsync;
 import org.jnosql.artemis.reflection.ClassRepresentation;
@@ -30,7 +30,7 @@ import java.lang.reflect.ParameterizedType;
 import java.util.function.Consumer;
 
 /**
- * Proxy handle to generate {@link CrudRepositoryAsync}
+ * Proxy handle to generate {@link RepositoryAsync}
  *
  * @param <T> the type
  */
@@ -45,7 +45,7 @@ class ColumnCrudRepositoryAsyncProxy<T> implements InvocationHandler {
 
     private final ColumnTemplateAsync repository;
 
-    private final ColumnCrudRepositoryAsync crudRepository;
+    private final ColumnRepositoryAsync crudRepository;
 
     private final ClassRepresentation classRepresentation;
 
@@ -56,7 +56,7 @@ class ColumnCrudRepositoryAsyncProxy<T> implements InvocationHandler {
 
     ColumnCrudRepositoryAsyncProxy(ColumnTemplateAsync repository, ClassRepresentations classRepresentations, Class<?> repositoryType) {
         this.repository = repository;
-        this.crudRepository = new ColumnCrudRepositoryAsync(repository);
+        this.crudRepository = new ColumnRepositoryAsync(repository);
         this.typeClass = Class.class.cast(ParameterizedType.class.cast(repositoryType.getGenericInterfaces()[0])
                 .getActualTypeArguments()[0]);
         this.classRepresentation = classRepresentations.get(typeClass);
@@ -105,11 +105,11 @@ class ColumnCrudRepositoryAsyncProxy<T> implements InvocationHandler {
     }
 
 
-    class ColumnCrudRepositoryAsync extends  AbstractColumnCrudRepositoryAsync implements CrudRepositoryAsync {
+    class ColumnRepositoryAsync extends AbstractColumnRepositoryAsync implements RepositoryAsync {
 
         private final ColumnTemplateAsync repository;
 
-        ColumnCrudRepositoryAsync(ColumnTemplateAsync repository) {
+        ColumnRepositoryAsync(ColumnTemplateAsync repository) {
             this.repository = repository;
         }
