+++ artemis-core/src/main/java/org/jnosql/artemis/document/query/DocumentCrudRepositoryProxy.java
@@ -16,7 +16,7 @@
 package org.jnosql.artemis.document.query;
 
 
-import org.jnosql.artemis.CrudRepository;
+import org.jnosql.artemis.Repository;
 import org.jnosql.artemis.document.DocumentTemplate;
 import org.jnosql.artemis.reflection.ClassRepresentation;
 import org.jnosql.artemis.reflection.ClassRepresentations;
@@ -30,7 +30,7 @@ import java.lang.reflect.ParameterizedType;
 
 
 /**
- * Proxy handle to generate {@link CrudRepository}
+ * Proxy handle to generate {@link Repository}
  *
  * @param <T> the type
  */
@@ -41,7 +41,7 @@ class DocumentCrudRepositoryProxy<T> implements InvocationHandler {
     private final DocumentTemplate repository;
 
 
-    private final DocumentCrudRepository crudRepository;
+    private final DocumentRepository crudRepository;
 
     private final ClassRepresentation classRepresentation;
 
@@ -52,7 +52,7 @@ class DocumentCrudRepositoryProxy<T> implements InvocationHandler {
 
     DocumentCrudRepositoryProxy(DocumentTemplate repository, ClassRepresentations classRepresentations, Class<?> repositoryType) {
         this.repository = repository;
-        this.crudRepository = new DocumentCrudRepository(repository);
+        this.crudRepository = new DocumentRepository(repository);
         this.typeClass = Class.class.cast(ParameterizedType.class.cast(repositoryType.getGenericInterfaces()[0])
                 .getActualTypeArguments()[0]);
         this.classRepresentation = classRepresentations.get(typeClass);
@@ -84,11 +84,11 @@ class DocumentCrudRepositoryProxy<T> implements InvocationHandler {
     }
 
 
-    class DocumentCrudRepository extends AbstractDocumentCrudRepository implements CrudRepository {
+    class DocumentRepository extends AbstractDocumentRepository implements Repository {
 
         private final DocumentTemplate repository;
 
-        DocumentCrudRepository(DocumentTemplate repository) {
+        DocumentRepository(DocumentTemplate repository) {
             this.repository = repository;
         }
