+++ artemis-core/src/test/java/org/jnosql/artemis/PersonRepositoryAsync.java
@@ -18,5 +18,5 @@ package org.jnosql.artemis;
 
 import org.jnosql.artemis.model.Person;
 
-public interface PersonRepositoryAsync extends CrudRepositoryAsync<Person> {
+public interface PersonRepositoryAsync extends RepositoryAsync<Person> {
 }
